gestion-de-donn-es-open-street-map-master/requirements.txt
pycairo==1.18.1
postgis==1.0.4
psycopg2==2.8.4
